Stub any_instance of Interceptor
Add default stub to set env to ‘test’ so that you only have to override
that if you want it to be something different.

This fixes a failure in the tests I was experiencing because
test_subject_prefix_in_production couldn’t stub env until the
interceptor was instantiated, but the initializer tests for production?
in the sanitize_forward_emails_to method called in the initializer.